+++ lib/gisele/language/ast/node.rb
@@ -34,6 +34,11 @@ module Gisele
           self[1..-1]
         end
 
+        # Returns the associated ast_module
+        def ast_module
+          AST::Helpers.send(:ast_module, self)
+        end
+
         # Applies copy-and-transform to this node.
         #
         # Example:
@@ -57,8 +62,29 @@ module Gisele
           AST.node(super, markers.dup)
         end
 
+        # Returns a label for this AST node
+        def label
+          ""
+        end
+
+        # Returns attributes to use for dot printing
+        def dot_attributes
+          attrs = Language::DOT_ATTRIBUTES[rule_name.to_s] || {}
+          attrs.merge(:label => label)
+        end
+
       end # module Node
     end # module AST
   end # module Language
 end # module Gisele
 require_relative 'unit'
+require_relative 'task_call_st'
+require_relative 'while_st'
+require_relative 'if_st'
+require_relative 'elsif_clause'
+require_relative 'when_clause'
+require_relative 'bool_expr'
+require_relative 'bool_and'
+require_relative 'bool_or'
+require_relative 'bool_not'
+require_relative 'var_ref'